Create the welcome screen for PacMan game
This commit adds the welcome screen that appears when
the player starts the game. The screen contains an image of the logo,
a text message, and a button like an image, which the player will press
to start the game and move to the next step.
While this class is not essential, it will make the game look more professional.