Improving code quality of the game
The commit adds more documentation and cleans up the code.
This is achieved by writing docstrings and the type annotations.
This is being done to improve the code quality so that it is more
accessible by other coders.
While these changes do not affect the functionality, it improves
the structure and makes our code look more professional.