Complete README.md for PacMan
This commit updates the empty README.md file. 
This new README contains everything that is needed to have on this page
to finalize the project repository section of the entire assignment.
This page is being updated so that in future people can 
easily understand what this project is all about.